servicios_clinica/requirements.txt
Flask
flask-marshmallow
Flask-RESTful
Flask-SQLAlchemy
SQLAlchemy
marshmallow-sqlalchemy
